Ajuste na classe de salario maior que 500
Removi o (byte) porque é desnecessário pro código rodar

+++ RHSystem_webapi/Program.cs
@@ -155,7 +155,7 @@ namespace RHSystem_webapi
 
 			// listar folha de pagamento com os salários maiores que 500 reais.
 			app.MapGet("/salariomaior", (Database basedeDados, int salario) => {
-			 	var result = basedeDados.Folha.Where(((byte)Folha.salario )>= 500);
+			 	var result = basedeDados.Folha.Where((Folha.salario )>= 500);
 				return result;
 				
 			 });
